Revert my work on async preferences
Working on this I realized getting each preference asynchronously is too
much. This would be a pain to maintain. I stored my previous work in a
local branch.
It would be better to get the preferences once at startup and keep an
in-memory object to read and write synchronously to. This object could
then be asynchronously copied to the chrome storage. We just have to
make sure to have written all preferences before finally closing the
application